// db/model_generated.ts
    export type id = string;
// ...
export namespace toursFields {
    export type id = string;
    export type location = string;
    export type tour_title = string;
    export type tour_category = string;
    export type tour_description = string | null;
    export type price = number;
    export type currency = string;
    export type img = Array<string> | null;

}

export interface tours {
    id: toursFields.id;
    location: toursFields.location;
    tour_title: toursFields.tour_title;
    tour_category: toursFields.tour_category;
    tour_description: toursFields.tour_description;
    price: toursFields.price;
    currency: toursFields.currency;
    img: toursFields.img;

}